V40: full-stack audit, CI remediation, composition evolution handoff
CI remediation (207 → 0 clippy errors):
- deny.toml: migrated to cargo-deny 0.19+ schema, added CC0-1.0, wildcards warn
- Library clippy: refactored neural.rs (228→3 helpers <100 LOC each), #[expect]
  for GPU casts, map_unwrap_or, collapsible_if
- Test clippy: #[allow] on 26+ test modules, real code quality fixes (redundant
  clones, collect removal, hypot, mul_add, extension checks)
- exp054: added missing license.workspace = true

Documentation and composition:
- 10 primal gaps documented (GAP-09 nest_atomic stubs, GAP-10 game.* identity)
- Test counts reconciled: 733 workspace (605 lib + 102 targets + 26 forge)
- load_baseline_f64: dead code → 4 unit tests
- Provenance verified: 4b683e3e valid, 13 semantic labels documented
- All docs updated to V40: README, CONTEXT, CHANGELOG, whitePaper, experiments
- V40 handoff crafted for primal + spring teams at infra/wateringHole/

Zero: clippy warnings, fmt diffs, unsafe, #[allow] in production, TODO/FIXME,
dead scripts, orphaned experiments. deny clean, baselines no drift.

Made-with: Cursor

+++ barracuda/src/game/rpgpt/scene.rs
@@ -275,6 +275,7 @@ pub struct NarrationEntry {
 }
 
 #[cfg(test)]
+#[allow(clippy::unwrap_used, clippy::expect_used)]
 mod tests {
     use super::*;
 
@@ -382,7 +383,10 @@ mod tests {
             actions_remaining: 3,
         };
         assert_eq!(grid.entities.len(), 2);
-        assert_eq!(grid.width * grid.height, grid.terrain.len() as u32);
+        assert_eq!(
+            (grid.width as usize) * (grid.height as usize),
+            grid.terrain.len()
+        );
     }
 
     #[test]
